% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{proportions}
\alias{proportions}
\title{proportions}
\usage{
proportions(KeithClass, Year.min, Year.max, sqmetres, TFI)
}
\arguments{
\item{KeithClass}{Keith Class that is desired (see provided table for list)}

\item{Year.min}{Minimum year of fire history information}

\item{Year.max}{Maximum year of fire history information}

\item{sqmetres}{Resolution of input pixels (square metres)}

\item{TFI}{TFI Thresholds}
}
\value{
Matrix of percent of total area that is within, below or above TFI thresholds for a given Keith Class
}
\description{
Function allows the tolerable fire interval proportions to be calculated above, below and within predetermined thresholds
}
