% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squaring stuf}
\usage{
square(x)
}
\arguments{
\item{x}{Describing the Arguments (i.e. what x is - Input Vector)}
}
\value{
Describes the output
}
\description{
Squares the input
}
\examples{
square(2:5)
}
\seealso{
link to another code
}
